resolve odataidresolver and apihandlers
