+++ src/Microsoft.AspNet.OData.Shared/ODataAPIHandler.cs
@@ -5,7 +5,12 @@
 // </copyright>
 //------------------------------------------------------------------------------
 
+using System;
+using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using Microsoft.OData.Edm;
 
 namespace Microsoft.AspNet.OData
 {
@@ -13,7 +18,7 @@ namespace Microsoft.AspNet.OData
     /// Handler Class to handle users methods for create, delete and update.
     /// This is the handler for data modification where there is a CLR type.
     /// </summary>
-    internal abstract class ODataAPIHandler<TStructuralType>: IODataAPIHandler where TStructuralType : class
+    public abstract class ODataAPIHandler<TStructuralType>: IODataAPIHandler where TStructuralType : class
     {
         /// <summary>
         /// TryCreate method to create a new object.
@@ -50,5 +55,90 @@ namespace Microsoft.AspNet.OData
         /// <param name="navigationPropertyName">The name of the navigation property for the handler</param>
         /// <returns>The type of Nested ODataAPIHandler</returns>
         public abstract IODataAPIHandler GetNestedHandler(TStructuralType parent, string navigationPropertyName);
+
+        /// <summary>
+        /// Apply OdataId for a resource with OdataID container
+        /// </summary>
+        /// <param name="resource">resource to apply odata id on</param>
+        /// <param name="model">The model.</param>
+        public virtual void UpdateLinkedObjects(TStructuralType resource, IEdmModel model)
+        {
+            if (resource != null && model != null)
+            {
+                CheckAndApplyODataId(resource);
+            }
+        }
+
+        private void CheckAndApplyODataId(object obj)
+        {
+            Type type = obj.GetType();
+
+            PropertyInfo property = type.GetProperties().FirstOrDefault(s => s.PropertyType == typeof(IODataIdContainer));
+
+            if (property != null && property.GetValue(obj) is IODataIdContainer container && container != null)
+            {
+                TStructuralType res = ApplyODataIdOnContainer(container);
+
+                foreach (PropertyInfo prop in type.GetProperties())
+                {
+                    object resVal = prop.GetValue(res);
+
+                    if (resVal != null)
+                    {
+                        prop.SetValue(obj, resVal);
+                    }
+                }
+            }
+            else
+            {
+                foreach (PropertyInfo prop in type.GetProperties().Where(p => !p.PropertyType.IsPrimitive))
+                {
+                    object propVal = prop.GetValue(obj);
+                    if (propVal == null)
+                    {
+                        continue;
+                    }
+
+                    if (propVal is IEnumerable lst)
+                    {
+                        foreach (object item in lst)
+                        {
+                            if (item.GetType().IsPrimitive)
+                            {
+                                break;
+                            }
+
+                            CheckAndApplyODataId(item);
+                        }
+                    }
+                    else
+                    {
+                        CheckAndApplyODataId(propVal);
+                    }
+                }
+            }
+
+        }
+
+        private TStructuralType ApplyODataIdOnContainer(IODataIdContainer container)
+        {
+            TStructuralType returnedObject;
+            string error;
+            if (TryGet(null, out returnedObject, out error) == ODataAPIResponseStatus.Success)
+            {
+                return returnedObject;
+            }
+            else 
+            {
+                if (TryCreate(null, out returnedObject, out error) == ODataAPIResponseStatus.Success)
+                {
+                    return returnedObject;
+                }
+                else 
+                {
+                    return null;
+                }
+            }
+        }
     }
 }
